pyproject.toml
[project]
name = "dog-identifier-app"
version = "0.1.0"
description = "A prototype web app to identify dogs by comparing images to known embeddings."
authors = [{ name = "Mike Charles", email = "serverlessmike@gmail.com" }]
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pillow",
  "numpy",
  "scipy",
  "transformers",
  "torch",
  "python-multipart"
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pylint",
    "black",
    "mypy"
]

[tool.pylint]
max-line-length = 100
indent-string = "\t"
indent-after-paren = 1

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = ">=3.9"

